Moved notifications related components to its own directory
