+++ apps/admin-x-activitypub/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@tryghost/admin-x-activitypub",
-  "version": "0.6.74",
+  "version": "0.6.75",
   "license": "MIT",
   "repository": {
     "type": "git",
